requirements.txt
openai
anthropic
aiogram
python-dotenv
